ocean-backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
anyio==3.7.0
arrow==1.2.3
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
billiard==4.1.0
binaryornot==0.4.4
celery==5.3.1
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cookiecutter==2.1.1
fastapi==0.98.0
fastapi-cache2==0.2.1
flower==2.0.0
frozenlist==1.3.3
h11==0.14.0
humanize==4.7.0
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
kombu==5.3.1
MarkupSafe==2.1.3
multidict==6.0.4
pendulum==2.1.2
prometheus-client==0.17.0
prompt-toolkit==3.0.38
pydantic==1.10.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0
redis==4.6.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.17
starlette==0.27.0
text-unidecode==1.3
tornado==6.3.2
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
uvicorn==0.22.0
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
